+++ domain/dto/org_dto.go
@@ -2,18 +2,18 @@ package dto
 
 type CreateOrgUnitRequest struct {
 	Name     string `json:"name" binding:"required"`
-	ParentID *uint  `json:"parent_id"`
+	ParentID *uint64  `json:"parent_id"`
 }
 
 type UpdateOrgUnitRequest struct {
 	Name     string `json:"name"`
-	ParentID *uint  `json:"parent_id"`
+	ParentID *uint64  `json:"parent_id"`
 }
 
 type OrgUnitResponse struct {
-	ID       uint   `json:"id"`
+	ID       uint64   `json:"id"`
 	Name     string `json:"name"`
-	ParentID *uint  `json:"parent_id"`
+	ParentID *uint64  `json:"parent_id"`
 	Path     string `json:"path"`
 	Level    int    `json:"level"`
 }
