+++ domain/dto/user_dto.go
@@ -5,8 +5,8 @@ type CreateUserRequest struct {
 	Name     string `json:"name" binding:"required"`
 	Email    string `json:"email" binding:"required,email"`
 	Password string `json:"password" binding:"required,min=6"`
-	RoleIDs  []uint `json:"role_ids"`
-	OrgUnitID uint  `json:"org_unit_id"`
+	RoleIDs  []uint64 `json:"role_ids"`
+	OrgUnitID uint64  `json:"org_unit_id"`
 }
 
 // UpdateUserRequest 更新用户请求（密码与角色为可选）
@@ -14,22 +14,22 @@ type UpdateUserRequest struct {
 	Name     string `json:"name"`
 	Email    string `json:"email" binding:"omitempty,email"`
 	Password string `json:"password" binding:"omitempty,min=6"`
-	RoleIDs  []uint `json:"role_ids"`
-	OrgUnitID uint  `json:"org_unit_id"`
+	RoleIDs  []uint64 `json:"role_ids"`
+	OrgUnitID uint64  `json:"org_unit_id"`
 }
 
 // UserResponse 用户响应
 type UserResponse struct {
-	ID      uint     `json:"id"`
+	ID      uint64     `json:"id"`
 	Name    string   `json:"name"`
 	Email   string   `json:"email"`
-	RoleIDs []uint   `json:"role_ids"`
+	RoleIDs []uint64   `json:"role_ids"`
 	Roles   []string `json:"roles"`
 }
 
 // ProfileResponse 个人信息响应
 type ProfileResponse struct {
-	ID        uint   `json:"id"`
+	ID        uint64   `json:"id"`
 	Name      string `json:"name"`
 	Email     string `json:"email"`
 	CreatedAt string `json:"created_at"`
