+++ domain/services/role_service.go
@@ -5,19 +5,19 @@ import (
 )
 
 type RoleService interface {
-	CreateRole(role *entity.Role, operatorID uint) error
-	UpdateRole(role *entity.Role, operatorID uint) error
-	DeleteRole(id uint, operatorID uint) error
-	GetRoleByID(id uint) (*entity.Role, error)
+	CreateRole(role *entity.Role, operatorID uint64) error
+	UpdateRole(role *entity.Role, operatorID uint64) error
+	DeleteRole(id uint64, operatorID uint64) error
+	GetRoleByID(id uint64) (*entity.Role, error)
 	GetAllRoles() ([]entity.Role, error)
-	BindResource(roleID, resourceID uint, isWrite bool, operatorID uint) error
-	UnbindResource(roleID, resourceID uint, operatorID uint) error
-	BindOrgScope(roleID, orgUnitID uint, includeDescendants bool, operatorID uint) error
-	UnbindOrgScope(roleID, orgUnitID uint, operatorID uint) error
-	AssignRoleToUser(userID, roleID, orgUnitID uint, operatorID uint) error
-	RevokeRoleFromUser(userID, roleID, orgUnitID uint, operatorID uint) error
-	GetRoleResources(roleID uint) ([]entity.RoleResource, error)
-	BindMenu(roleID, menuID uint, operatorID uint) error
-	UnbindMenu(roleID, menuID uint, operatorID uint) error
-	GetRoleMenus(roleID uint) ([]entity.RoleMenu, error)
+	BindResource(roleID, resourceID uint64, isWrite bool, operatorID uint64) error
+	UnbindResource(roleID, resourceID uint64, operatorID uint64) error
+	BindOrgScope(roleID, orgUnitID uint64, includeDescendants bool, operatorID uint64) error
+	UnbindOrgScope(roleID, orgUnitID uint64, operatorID uint64) error
+	AssignRoleToUser(userID, roleID, orgUnitID uint64, operatorID uint64) error
+	RevokeRoleFromUser(userID, roleID, orgUnitID uint64, operatorID uint64) error
+	GetRoleResources(roleID uint64) ([]entity.RoleResource, error)
+	BindMenu(roleID, menuID uint64, operatorID uint64) error
+	UnbindMenu(roleID, menuID uint64, operatorID uint64) error
+	GetRoleMenus(roleID uint64) ([]entity.RoleMenu, error)
 }
